CMakeLists refactor

+++ sprint3/problems/find_return/solution/src/game/game.h
@@ -64,7 +64,7 @@ private:
     double       defaultSpeed_;
     uint32_t     defaultBagCapacity_{3};
     bool         randomSpawn_ = false;
-    double       randomGeneratorPeriod_;
+    uint32_t     randomGeneratorPeriod_;
     double       randomGeneratorProbability_;
 };
